Load/postgres/main/requirements.txt
confluent-kafka
psycopg2-binary
pandas
